+++ src/ui-test/all.ts
@@ -1,7 +1,5 @@
-import { basicTests } from "./basic-test";
 import { editorTests } from './editor';
 
 describe('AtlasMap UI tests', () => {
 	editorTests();
-	//basicTests();
 });